# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Meanwhile","description":"Meanwhile - Intelligent Market Catch-Up Engine (CODE by Groww 2026 Submission)","version":"1.0.0"},"paths":{"/api/v1/auth/demo-login":{"post":{"tags":["Authentication"],"summary":"Demo Login","description":"Provides a 1-click instant login for hackathon evaluation.","operationId":"demo_login_api_v1_auth_demo_login_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}}}}},"/api/v1/auth/signup":{"post":{"tags":["Authentication"],"summary":"Signup","operationId":"signup_api_v1_auth_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/watchlists":{"get":{"tags":["Watchlists"],"summary":"Get Watchlists","operationId":"get_watchlists_api_v1_watchlists_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/WatchlistResponse"},"type":"array","title":"Response Get Watchlists Api V1 Watchlists Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["Watchlists"],"summary":"Create Watchlist","operationId":"create_watchlist_api_v1_watchlists_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchlistCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchlistResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/watchlists/{watchlist_id}":{"get":{"tags":["Watchlists"],"summary":"Get Watchlist","operationId":"get_watchlist_api_v1_watchlists__watchlist_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"watchlist_id","in":"path","required":true,"schema":{"type":"integer","title":"Watchlist Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchlistResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Watchlists"],"summary":"Update Watchlist","operationId":"update_watchlist_api_v1_watchlists__watchlist_id__patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"watchlist_id","in":"path","required":true,"schema":{"type":"integer","title":"Watchlist Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchlistUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchlistResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Watchlists"],"summary":"Delete Watchlist","operationId":"delete_watchlist_api_v1_watchlists__watchlist_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"watchlist_id","in":"path","required":true,"schema":{"type":"integer","title":"Watchlist Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/watchlists/{watchlist_id}/items":{"post":{"tags":["Watchlists"],"summary":"Add Item","operationId":"add_item_api_v1_watchlists__watchlist_id__items_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"watchlist_id","in":"path","required":true,"schema":{"type":"integer","title":"Watchlist Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchlistItemCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchlistItemResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/watchlists/{watchlist_id}/items/{symbol}":{"delete":{"tags":["Watchlists"],"summary":"Remove Item","operationId":"remove_item_api_v1_watchlists__watchlist_id__items__symbol__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"watchlist_id","in":"path","required":true,"schema":{"type":"integer","title":"Watchlist Id"}},{"name":"symbol","in":"path","required":true,"schema":{"type":"string","title":"Symbol"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/attention/since-last-check":{"get":{"tags":["Attention Engine"],"summary":"Get Since Last Check Feed","description":"Main catch-up endpoint returning ranked market changes since the user's last check,\ncategorized by severity (HIGH_ATTENTION, MEANINGFUL, WATCH, NORMAL).","operationId":"get_since_last_check_feed_api_v1_attention_since_last_check_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"watchlist_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"description":"Optional filter by watchlist ID","title":"Watchlist Id"},"description":"Optional filter by watchlist ID"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SinceLastCheckResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/attention/acknowledge-seen":{"post":{"tags":["Attention Engine"],"summary":"Acknowledge Seen","description":"Marks current attention items as seen, updating last_seen_at timestamp to NOW.","operationId":"acknowledge_seen_api_v1_attention_acknowledge_seen_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"watchlist_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Watchlist Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/attention/why/{symbol}":{"get":{"tags":["Attention Engine"],"summary":"Get Why Breakdown","description":"Returns full transparent mathematical score breakdown and bullet reasons for a specific stock.","operationId":"get_why_breakdown_api_v1_attention_why__symbol__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"symbol","in":"path","required":true,"schema":{"type":"string","title":"Symbol"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeEventResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/market/quote/{symbol}":{"get":{"tags":["Market Data"],"summary":"Get Quote","operationId":"get_quote_api_v1_market_quote__symbol__get","parameters":[{"name":"symbol","in":"path","required":true,"schema":{"type":"string","title":"Symbol"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MarketSnapshotResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/market/benchmark":{"get":{"tags":["Market Data"],"summary":"Get Benchmark","operationId":"get_benchmark_api_v1_market_benchmark_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MarketSnapshotResponse"}}}}}}},"/api/v1/market/symbols":{"get":{"tags":["Market Data"],"summary":"Get Available Symbols","operationId":"get_available_symbols_api_v1_market_symbols_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"type":"string"},"type":"array","title":"Response Get Available Symbols Api V1 Market Symbols Get"}}}}}}},"/api/v1/simulation/set-scenario":{"post":{"tags":["Market Simulation Control"],"summary":"Set Simulation Scenario","description":"Evaluator tool: switches the active mock market state live\n(NORMAL, VOLUME_SPIKE, MARKET_RALLY, EARNINGS_BREAKOUT, STALE_DATA, CONFLICTING_DATA).","operationId":"set_simulation_scenario_api_v1_simulation_set_scenario_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScenarioRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/simulation/reset":{"post":{"tags":["Market Simulation Control"],"summary":"Reset Simulation Scenario","description":"Evaluator tool: resets market scenarios back to normal.","operationId":"reset_simulation_scenario_api_v1_simulation_reset_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AttentionSummary":{"properties":{"total_watched":{"type":"integer","title":"Total Watched"},"high_attention_count":{"type":"integer","title":"High Attention Count"},"meaningful_count":{"type":"integer","title":"Meaningful Count"},"watch_count":{"type":"integer","title":"Watch Count"},"normal_count":{"type":"integer","title":"Normal Count"},"last_checked_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Checked At"},"elapsed_time_formatted":{"type":"string","title":"Elapsed Time Formatted"},"has_material_changes":{"type":"boolean","title":"Has Material Changes","default":true},"attention_budget":{"type":"integer","title":"Attention Budget","default":5}},"type":"object","required":["total_watched","high_attention_count","meaningful_count","watch_count","normal_count","elapsed_time_formatted"],"title":"AttentionSummary"},"ChangeEventResponse":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"symbol":{"type":"string","title":"Symbol"},"watchlist_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Watchlist Id"},"detected_at":{"type":"string","format":"date-time","title":"Detected At"},"previous_price":{"type":"number","title":"Previous Price"},"current_price":{"type":"number","title":"Current Price"},"price_change_percent":{"type":"number","title":"Price Change Percent"},"volume_change_percent":{"type":"number","title":"Volume Change Percent"},"volume_multiplier":{"type":"number","title":"Volume Multiplier","default":1.0},"market_relative_change":{"type":"number","title":"Market Relative Change"},"market_outperformance_label":{"type":"string","title":"Market Outperformance Label","default":""},"score":{"type":"number","title":"Score"},"severity":{"$ref":"#/components/schemas/SeverityLevel"},"reasons":{"items":{"type":"string"},"type":"array","title":"Reasons"},"score_breakdown":{"$ref":"#/components/schemas/ScoreBreakdownSchema"},"data_quality":{"$ref":"#/components/schemas/DataQuality"},"timeline":{"items":{"$ref":"#/components/schemas/PriceTimelinePointSchema"},"type":"array","title":"Timeline","default":[]},"is_within_attention_budget":{"type":"boolean","title":"Is Within Attention Budget","default":true}},"type":"object","required":["symbol","detected_at","previous_price","current_price","price_change_percent","volume_change_percent","market_relative_change","score","severity","reasons","score_breakdown","data_quality"],"title":"ChangeEventResponse"},"DataQuality":{"type":"string","enum":["FRESH","DELAYED","STALE","CONFLICTING"],"title":"DataQuality"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"MarketRegimeSchema":{"properties":{"benchmark_symbol":{"type":"string","title":"Benchmark Symbol","default":"NIFTY50"},"benchmark_change_pct":{"type":"number","title":"Benchmark Change Pct","default":0.5},"advancing_pct":{"type":"number","title":"Advancing Pct","default":62.0},"advancing_count":{"type":"integer","title":"Advancing Count","default":31},"declining_count":{"type":"integer","title":"Declining Count","default":19},"volatility_regime":{"type":"string","title":"Volatility Regime","default":"Normal"},"regime_label":{"type":"string","title":"Regime Label","default":"Stock-specific movement detected"}},"type":"object","title":"MarketRegimeSchema"},"MarketSnapshotResponse":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"symbol":{"type":"string","title":"Symbol"},"price":{"type":"number","title":"Price"},"change":{"type":"number","title":"Change"},"change_percent":{"type":"number","title":"Change Percent"},"volume":{"type":"number","title":"Volume"},"avg_volume_20d":{"type":"number","title":"Avg Volume 20D"},"volatility_30d":{"type":"number","title":"Volatility 30D"},"benchmark_symbol":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Benchmark Symbol","default":"NIFTY50"},"benchmark_price_change_pct":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Benchmark Price Change Pct","default":0.0},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"source":{"type":"string","title":"Source"},"data_quality":{"$ref":"#/components/schemas/DataQuality"},"timeline":{"items":{"$ref":"#/components/schemas/PriceTimelinePointSchema"},"type":"array","title":"Timeline","default":[]}},"type":"object","required":["symbol","price","change","change_percent","volume","avg_volume_20d","volatility_30d","timestamp","source","data_quality"],"title":"MarketSnapshotResponse"},"PriceTimelinePointSchema":{"properties":{"time":{"type":"string","title":"Time"},"price":{"type":"number","title":"Price"}},"type":"object","required":["time","price"],"title":"PriceTimelinePointSchema"},"ScenarioRequest":{"properties":{"scenario":{"$ref":"#/components/schemas/SimulationScenario"}},"type":"object","required":["scenario"],"title":"ScenarioRequest"},"ScoreBreakdownSchema":{"properties":{"price_score":{"type":"number","title":"Price Score","default":0.0},"volume_score":{"type":"number","title":"Volume Score","default":0.0},"market_relative_score":{"type":"number","title":"Market Relative Score","default":0.0},"volatility_score":{"type":"number","title":"Volatility Score","default":0.0},"event_score":{"type":"number","title":"Event Score","default":0.0},"data_quality_penalty":{"type":"number","title":"Data Quality Penalty","default":0.0},"total_score":{"type":"number","title":"Total Score","default":0.0}},"type":"object","title":"ScoreBreakdownSchema"},"SeverityLevel":{"type":"string","enum":["NORMAL","WATCH","MEANINGFUL","HIGH_ATTENTION"],"title":"SeverityLevel"},"SimulationScenario":{"type":"string","enum":["NORMAL","VOLUME_SPIKE","MARKET_RALLY","EARNINGS_BREAKOUT","STALE_DATA","CONFLICTING_DATA"],"title":"SimulationScenario"},"SinceLastCheckResponse":{"properties":{"summary":{"$ref":"#/components/schemas/AttentionSummary"},"market_regime":{"$ref":"#/components/schemas/MarketRegimeSchema"},"top_events":{"items":{"$ref":"#/components/schemas/ChangeEventResponse"},"type":"array","title":"Top Events"},"other_events":{"items":{"$ref":"#/components/schemas/ChangeEventResponse"},"type":"array","title":"Other Events"},"events":{"items":{"$ref":"#/components/schemas/ChangeEventResponse"},"type":"array","title":"Events"},"benchmark_symbol":{"type":"string","title":"Benchmark Symbol","default":"NIFTY50"},"benchmark_change_pct":{"type":"number","title":"Benchmark Change Pct","default":0.0}},"type":"object","required":["summary","market_regime","top_events","other_events","events"],"title":"SinceLastCheckResponse"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"user":{"$ref":"#/components/schemas/UserResponse"}},"type":"object","required":["access_token","user"],"title":"Token"},"UserCreate":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"name":{"type":"string","title":"Name","default":"Investor"}},"type":"object","required":["email","password"],"title":"UserCreate"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"email":{"type":"string","title":"Email"},"name":{"type":"string","title":"Name"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","email","name","created_at"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WatchlistCreate":{"properties":{"name":{"type":"string","maxLength":100,"minLength":1,"title":"Name"}},"type":"object","required":["name"],"title":"WatchlistCreate"},"WatchlistItemCreate":{"properties":{"symbol":{"type":"string","title":"Symbol","description":"Stock ticker symbol, e.g. INFY, RELIANCE"}},"type":"object","required":["symbol"],"title":"WatchlistItemCreate"},"WatchlistItemResponse":{"properties":{"id":{"type":"integer","title":"Id"},"watchlist_id":{"type":"integer","title":"Watchlist Id"},"symbol":{"type":"string","title":"Symbol"},"added_at":{"type":"string","format":"date-time","title":"Added At"},"last_seen_at":{"type":"string","format":"date-time","title":"Last Seen At"},"latest_snapshot":{"anyOf":[{"$ref":"#/components/schemas/MarketSnapshotResponse"},{"type":"null"}]}},"type":"object","required":["id","watchlist_id","symbol","added_at","last_seen_at"],"title":"WatchlistItemResponse"},"WatchlistResponse":{"properties":{"id":{"type":"integer","title":"Id"},"user_id":{"type":"integer","title":"User Id"},"name":{"type":"string","title":"Name"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"items":{"items":{"$ref":"#/components/schemas/WatchlistItemResponse"},"type":"array","title":"Items","default":[]}},"type":"object","required":["id","user_id","name","created_at"],"title":"WatchlistResponse"},"WatchlistUpdate":{"properties":{"name":{"anyOf":[{"type":"string","maxLength":100,"minLength":1},{"type":"null"}],"title":"Name"}},"type":"object","title":"WatchlistUpdate"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/auth/login"}}}}}}